♻️  Refactor : 데이터불러오는 api 함수들 useCallback 처리
