+++ src/components/comment/CommentList.jsx
@@ -1,5 +1,5 @@
-import { commentListAPI } from 'api/comment';
-import React, { useEffect, useState } from 'react';
+import { commentListAPI } from 'api/comment.api';
+import React, { useCallback, useEffect, useState } from 'react';
 import CommentCard from 'components/comment/CommentCard';
 import { Container } from 'components/comment/CommentList.style';
 
@@ -7,19 +7,19 @@ export default function CommentList({ postid, commentCount, update }) {
   const [comments, setComments] = useState(null);
 
   //댓글목록 불러오기
-  const commentList = () => {
+  const commentList = useCallback(async () => {
     commentListAPI(postid)
       .then((res) => {
-        setComments(commentlist);
+        setComments(res);
       })
       .catch(() => {
         alert('댓글 불러오기에 실패했습니다.');
       });
-  };
+  }, [postid]);
 
   useEffect(() => {
     commentList();
-  }, [commentCount]);
+  }, [commentCount, commentList]);
 
   const commentlist = () => {
     return comments.map((item) => {
