+++ src/pages/PostDetailPage/PostDetailPage.jsx
@@ -1,7 +1,7 @@
 import { ScrollContainer } from 'components/common/container/Container.style';
 import Header from 'components/common/header/Header';
 import PostCard from 'components/post/PostCard';
-import React, { useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 import { PostBox } from './PostDetailPage.style';
 import { postDetailAPI } from 'api/post.api';
@@ -15,14 +15,14 @@ export default function PostDetailPage() {
   const { postid } = useParams();
 
   //게시글정보 불러오기
-  const postDetail = async () => {
+  const postDetail = useCallback(async () => {
     try {
       const data = await postDetailAPI(postid);
       setPost(data);
     } catch (err) {
       alert('게시글 불러오기에 실패했습니다.');
     }
-  };
+  }, [postid]);
 
   useEffect(() => {
     postDetail();
